corrections from reviews
