+++ doc/source/code-examples/advancedexamples.rst
@@ -2011,10 +2011,9 @@ Multiple transpilation steps can be implemented using the :class:`qibo.transpile
     from qibo import gates
     from qibo.models import Circuit
     from qibo.transpiler.pipeline import Passes, assert_transpiling
-    from qibo.transpiler.abstract import NativeGates
     from qibo.transpiler.optimizer import Preprocessing
     from qibo.transpiler.router import ShortestPaths
-    from qibo.transpiler.unroller import DefaultUnroller
+    from qibo.transpiler.unroller import Unroller, NativeGates
     from qibo.transpiler.placer import Random
 
     # Define connectivity as nx.Graph
@@ -2040,7 +2039,7 @@ Multiple transpilation steps can be implemented using the :class:`qibo.transpile
     # Routing step
     custom_passes.append(ShortestPaths(connectivity=star_connectivity()))
     # Gate decomposition step
-    custom_passes.append(DefaultUnroller(native_gates=NativeGates.default()))
+    custom_passes.append(Unroller(native_gates=NativeGates.default()))
 
     # Define the general pipeline
     custom_pipeline = Passes(custom_passes, connectivity=star_connectivity(), native_gates=NativeGates.default())
@@ -2061,6 +2060,6 @@ Multiple transpilation steps can be implemented using the :class:`qibo.transpile
     )
 
 In this case circuits will first be transpiled to respect the 5-qubit star connectivity, with qubit 2 as the middle qubit. This will potentially add some SWAP gates.
-Then all gates will be converted to native. The :class:`qibo.transpiler.unroller.NativeGates` transpiler used in this example assumes Z, RZ, GPI2 or U3 as
+Then all gates will be converted to native. The :class:`qibo.transpiler.unroller.Unroller` transpiler used in this example assumes Z, RZ, GPI2 or U3 as
 the single-qubit native gates, and supports CZ and iSWAP as two-qubit natives. In this case we restricted the two-qubit gate set to CZ only.
 The final_layout contains the final logical-physical qubit mapping.
