+++ tests/test_transpiler_blocks.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pytest
 
 from qibo import Circuit, gates
