+++ tests/test_transpiler_decompositions.py
@@ -5,8 +5,9 @@ from qibo import gates
 from qibo.backends import NumpyBackend
 from qibo.models import Circuit
 from qibo.quantum_info.random_ensembles import random_unitary
-from qibo.transpiler.abstract import NativeGates
-from qibo.transpiler.unroller import assert_decomposition, translate_gate
+from qibo.transpiler.unroller import NativeGates, assert_decomposition, translate_gate
+
+default_natives = NativeGates.Z | NativeGates.RZ | NativeGates.M | NativeGates.I
 
 
 def assert_matrices_allclose(gate, natives, backend):
@@ -42,45 +43,45 @@ def assert_matrices_allclose(gate, natives, backend):
 @pytest.mark.parametrize("gatename", ["H", "X", "Y", "I"])
 def test_pauli_to_native(backend, gatename, natives):
     gate = getattr(gates, gatename)(0)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("natives", [NativeGates.U3, NativeGates.GPI2])
 @pytest.mark.parametrize("gatename", ["RX", "RY", "RZ"])
 def test_rotations_to_native(backend, gatename, natives):
     gate = getattr(gates, gatename)(0, theta=0.1)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("natives", [NativeGates.U3, NativeGates.GPI2])
 @pytest.mark.parametrize("gatename", ["S", "SDG", "T", "TDG", "SX"])
 def test_special_single_qubit_to_native(backend, gatename, natives):
     gate = getattr(gates, gatename)(0)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("natives", [NativeGates.U3, NativeGates.GPI2])
 def test_u1_to_native(backend, natives):
     gate = gates.U1(0, theta=0.5)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("natives", [NativeGates.U3, NativeGates.GPI2])
 def test_u2_to_native(backend, natives):
     gate = gates.U2(0, phi=0.1, lam=0.3)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("natives", [NativeGates.U3, NativeGates.GPI2])
 def test_u3_to_native(backend, natives):
     gate = gates.U3(0, theta=0.2, phi=0.1, lam=0.3)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("natives", [NativeGates.U3, NativeGates.GPI2])
 def test_gpi2_to_native(backend, natives):
     gate = gates.GPI2(0, phi=0.123)
-    assert_matrices_allclose(gate, natives=natives, backend=backend)
+    assert_matrices_allclose(gate, natives=natives | default_natives, backend=backend)
 
 
 @pytest.mark.parametrize("gatename", ["CNOT", "CZ", "SWAP", "iSWAP", "FSWAP"])
@@ -94,7 +95,7 @@ def test_gpi2_to_native(backend, natives):
 )
 def test_two_qubit_to_native(backend, gatename, natives_1q, natives_2q):
     gate = getattr(gates, gatename)(0, 1)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -108,7 +109,7 @@ def test_two_qubit_to_native(backend, gatename, natives_1q, natives_2q):
 @pytest.mark.parametrize("gatename", ["CRX", "CRY", "CRZ"])
 def test_controlled_rotations_to_native(backend, gatename, natives_1q, natives_2q):
     gate = getattr(gates, gatename)(0, 1, 0.3)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -121,7 +122,7 @@ def test_controlled_rotations_to_native(backend, gatename, natives_1q, natives_2
 )
 def test_cu1_to_native(backend, natives_1q, natives_2q):
     gate = gates.CU1(0, 1, theta=0.4)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -134,7 +135,7 @@ def test_cu1_to_native(backend, natives_1q, natives_2q):
 )
 def test_cu2_to_native(backend, natives_1q, natives_2q):
     gate = gates.CU2(0, 1, phi=0.2, lam=0.3)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -147,7 +148,7 @@ def test_cu2_to_native(backend, natives_1q, natives_2q):
 )
 def test_cu3_to_native(backend, natives_1q, natives_2q):
     gate = gates.CU3(0, 1, theta=0.2, phi=0.3, lam=0.4)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -160,7 +161,7 @@ def test_cu3_to_native(backend, natives_1q, natives_2q):
 )
 def test_fSim_to_native(backend, natives_1q, natives_2q):
     gate = gates.fSim(0, 1, theta=0.3, phi=0.1)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize("seed", [None, 10, np.random.default_rng(10)])
@@ -175,7 +176,7 @@ def test_fSim_to_native(backend, natives_1q, natives_2q):
 def test_GeneralizedfSim_to_native(backend, natives_1q, natives_2q, seed):
     unitary = random_unitary(2, seed=seed, backend=backend)
     gate = gates.GeneralizedfSim(0, 1, unitary, phi=0.1)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -189,7 +190,7 @@ def test_GeneralizedfSim_to_native(backend, natives_1q, natives_2q, seed):
 @pytest.mark.parametrize("gatename", ["RXX", "RZZ", "RYY"])
 def test_rnn_to_native(backend, gatename, natives_1q, natives_2q):
     gate = getattr(gates, gatename)(0, 1, theta=0.1)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize(
@@ -202,7 +203,7 @@ def test_rnn_to_native(backend, gatename, natives_1q, natives_2q):
 )
 def test_TOFFOLI_to_native(backend, natives_1q, natives_2q):
     gate = gates.TOFFOLI(0, 1, 2)
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 @pytest.mark.parametrize("seed", [None, 10, np.random.default_rng(10)])
@@ -220,7 +221,7 @@ def test_unitary_to_native(backend, nqubits, natives_1q, natives_2q, seed):
     # transform to SU(2^nqubits) form
     u = u / np.sqrt(np.linalg.det(u))
     gate = gates.Unitary(u, *range(nqubits))
-    assert_matrices_allclose(gate, natives_1q | natives_2q, backend)
+    assert_matrices_allclose(gate, natives_1q | natives_2q | default_natives, backend)
 
 
 def test_count_1q():
